1、BUG修改

+++ src/main/java/com/business/controller/pc/category/CategoryController.java
@@ -7,7 +7,9 @@ import com.business.controller.pc.category.dto.CategoryAddReqDTO;
 import com.business.controller.pc.category.dto.CategoryPageReqDTO;
 import com.business.controller.pc.category.dto.CategoryUpdateReqDTO;
 import com.business.controller.pc.category.vo.CategoryVO;
+import com.business.model.pojo.WallpaperInfo;
 import com.business.service.category.ICategoryService;
+import com.business.service.wallpaper.IWallpaperInfoService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -25,6 +27,9 @@ public class CategoryController {
 
     @Resource
     private ICategoryService categoryService;
+
+    @Resource
+    private IWallpaperInfoService wallpaperInfoService;
 
     @Operation(summary = "分页查询壁纸分类列表")
     @GetMapping("/pageList")
@@ -54,6 +59,11 @@ public class CategoryController {
     @Parameter(name = "id", description = "分类ID", required = true)
     @DeleteMapping("/deleteById")
     public ResultVO<Boolean> deleteById(@RequestParam("id") Long id){
+        //查询是否有壁纸跟该分类进行关联，如果有则不准删除
+        Long count = wallpaperInfoService.lambdaQuery().eq(WallpaperInfo::getCategoryId, id).count();
+        if (count > 0){
+            return ResultVO.error("该分类下有壁纸，不允许删除");
+        }
         categoryService.removeById(id);
         return ResultVO.success(true);
     }
